[Refactoring] Divide ssd.py file as per class functionality
